Abstracted getFifths to its own module
