+++ src/components/CircleOfFifths/utils/index.ts
@@ -1,1 +1,2 @@
+export { default as getFifths } from './getFifths';
 export { default as getModeOffset } from './getModeOffset';
